+++ Assets/Blockiverse/Scripts/Voxel/VoxelAssembly.cs
@@ -1,0 +1,7 @@
+namespace Blockiverse.Voxel
+{
+    public static class VoxelAssembly
+    {
+        public const string Name = "Blockiverse.Voxel";
+    }
+}