+++ Assets/Blockiverse/Scripts/WorldGen/WorldGenAssembly.cs
@@ -1,0 +1,7 @@
+namespace Blockiverse.WorldGen
+{
+    public static class WorldGenAssembly
+    {
+        public const string Name = "Blockiverse.WorldGen";
+    }
+}